tests/wpt/css-tests/cssom-view-1_dev/xhtml1/cssom-view-window-screen-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
	<title>CSS Test: cssom view window screen attribute</title>
	<link href="mailto:jingkewhu@gmail.com" rel="author" title="jingke" />
	<link href="http://www.w3.org/TR/cssom-view/#the-screen-interface" rel="help" />
	<meta content="dom" name="flags" />
	<meta content="window screen interface" name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
	</head>
	<body>
		<div id="myDiv"></div>
        <div id="log"></div>
		<script>
		/*test readonly*/
		test(function(){assert_readonly(window.screen, "availWidth");}, "Screen.availWidth is readonly");
		test(function(){assert_readonly(window.screen, "availHeight");}, "Screen.availHeight is readonly");
		test(function(){assert_readonly(window.screen, "width");}, "Screen.width is readonly");
		test(function(){assert_readonly(window.screen, "height");}, "Screen.height is readonly");
		test(function(){assert_readonly(window.screen, "colorDepth");}, "Screen.colorDepth is readonly");
		test(function(){assert_readonly(window.screen, "pixelDepth");}, "Screen.pixelDepth is readonly");


		test(function(){assert_true(window.screen.width &gt;= 0 &amp;&amp; window.screen.width &lt; 6000000);},
			"window.screen.width &gt;= 0 &amp;&amp; window.screen.width &lt; 6000000");
		test(function(){assert_true(window.screen.height &gt;= 0 &amp;&amp; window.screen.height &lt; 6000000);},
			"window.screen.height &gt;= 0 &amp;&amp; window.screen.height &lt; 6000000");
		test(function(){assert_true(window.screen.availWidth &gt;= 0 &amp;&amp; window.screen.availWidth &lt;= window.screen.width);},
			"window.screen.availWidth &gt;= 0 &amp;&amp; window.screen.availWidth &lt;= window.screen.width");
		test(function(){assert_true(window.screen.availHeight &gt;= 0 &amp;&amp; window.screen.availHeight &lt;= window.screen.height);},
			"window.screen.availHeight &gt;= 0 &amp;&amp; window.screen.availHeight &lt;= window.screen.height");
		test(function(){assert_true(window.screen.colorDepth == 0 || window.screen.colorDepth == 16 || window.screen.colorDepth == 24 || window.screen.colorDepth == 32);},
			"window.screen.colorDepth == 0 || window.screen.colorDepth == 16 || window.screen.colorDepth == 24 || window.screen.colorDepth == 32");
		test(function(){assert_equals(window.screen.pixelDepth, window.screen.colorDepth);},
			"window.screen.pixelDepth must return the value returned by window.screen.colorDepth");
		</script>

	

</body></html>
